requirements.txt
chainlit==0.7.700
python-dotenv==1.0.0
google-cloud-bigquery==3.14.1
google-auth==2.29.0
google-cloud-core==2.3.3
protobuf==4.24.4
pydantic==2.6.1
typing-extensions==4.9.0 